requirements.txt
colorama>=0.4.3
mysql-connector>=2.2.9
